requirements.txt
langchain
python-dotenv
langchain-openai
beautifulsoup4
langchain_community
langchain_ollama
faiss-cpu
bs4
langchain_groq
langchain_core
langchain_chroma
fastapi
uvicorn
langserve
sse_starlette
langchain_huggingface
sentence-transformers
streamlit